fix setup db for pytests
