+++ web/test/test_login.py
@@ -1,25 +1,9 @@
-import pytest
-from app.app import create_app, db
-from app.util import setup_db
-
 CORRECT_MAIL = "admin@stella-project.org"
 CORRECT_PASS = "pass"
 INCORRECT_MAIL = "anonymous@stella-project.org"
 MSG_GREETINGS = b"Hello, stella-admin!"
 MSG_LOGOUT = b"You have been logged out."
 MSG_INVALID = b"Invalid email or password."
-
-
-@pytest.fixture
-def client():
-    app = create_app("default")
-    app.config["WTF_CSRF_ENABLED"] = False
-    app.config["TESTING"] = True
-    with app.test_client() as client:
-        with app.app_context():
-
-            setup_db(db)
-        yield client
 
 
 def logout(client, email, password):
